Updated fixtures structure
